Aphasia/requirements.txt
numpy 
pandas  
seaborn 
matplotlib 
optuna 
scikit-learn
imblearn 
pyyaml 
xgboost
lightgbm